refactor(capcut-e2e): split visual contract validation
